refactor: move common test utilities to separate package
