+++ packages/nostr-fetch/package.json
@@ -45,12 +45,11 @@
   "dependencies": {
     "@nostr-fetch/kernel": "^0.9.0"
   },
+  "devDependencies": {
+    "@nostr-fetch/testutil": "^0.9.0",
+    "vitest-websocket-mock": "^0.1.3"
+  },
   "volta": {
     "node": "18.14.2"
-  },
-  "devDependencies": {
-    "@noble/hashes": "^1.3.1",
-    "nostr-tools": "^1.12.0",
-    "vitest-websocket-mock": "^0.1.3"
   }
 }
