+++ packages/testutil/README.md
@@ -1,0 +1,1 @@
+# Test Utilities for nostr-fetch
